+++ include/sycamore/desktop/shell/layer_shell/layer.h
@@ -11,11 +11,12 @@ struct sycamore_output;
 struct sycamore_layer {
     enum scene_descriptor_type scene_descriptor;    //must be first
     struct wlr_layer_surface_v1 *layer_surface;
+    struct wlr_scene_layer_surface_v1 *scene;
+
     bool mapped;
+
     bool linked;
     struct wl_list link;
-
-    struct wlr_scene_layer_surface_v1 *scene;
 
     struct wl_listener destroy;
     struct wl_listener map;
