+++ include/sycamore/desktop/shell/layer_shell/layer_shell.h
@@ -7,7 +7,6 @@
 
 struct sycamore_layer_shell {
     struct wlr_layer_shell_v1 *wlr_layer_shell;
-
     struct wl_listener new_layer_shell_surface;
 };
