+++ sycamore/desktop/shell/layer_shell/layer_shell.c
@@ -1,36 +1,10 @@
 #include <stdlib.h>
 #include <wlr/types/wlr_layer_shell_v1.h>
 #include <wlr/util/log.h>
-#include "sycamore/desktop/layer.h"
-#include "sycamore/desktop/shell/layer_shell.h"
+#include "sycamore/desktop/shell/layer_shell/layer.h"
+#include "sycamore/desktop/shell/layer_shell/layer_shell.h"
 #include "sycamore/output/output.h"
 #include "sycamore/server.h"
-
-static void handle_layer_map(struct wl_listener *listener, void *data) {
-    struct sycamore_layer *layer = wl_container_of(listener, layer, map);
-
-    layer_map(layer);
-}
-
-static void handle_layer_unmap(struct wl_listener *listener, void *data) {
-    struct sycamore_layer *layer = wl_container_of(listener, layer, unmap);
-
-    layer_unmap(layer);
-}
-
-static void handle_layer_destroy(struct wl_listener *listener, void *data) {
-    struct sycamore_layer *layer = wl_container_of(listener, layer, destroy);
-
-    layer->layer_surface = NULL;
-
-    layer_destroy(layer);
-}
-
-static void handle_layer_surface_commit(struct wl_listener *listener, void *data) {
-    struct sycamore_layer *layer = wl_container_of(listener, layer, surface_commit);
-
-    layer_surface_commit(layer);
-}
 
 static void handle_new_layer_shell_surface(struct wl_listener *listener, void *data) {
     struct sycamore_layer_shell *layer_shell =
@@ -54,15 +28,6 @@ static void handle_new_layer_shell_surface(struct wl_listener *listener, void *d
     struct sycamore_output *output = layer->output;
     wl_list_insert(&output->layers[layer_type], &layer->link);
     layer->linked = true;
-
-    layer->map.notify = handle_layer_map;
-    wl_signal_add(&layer_surface->events.map, &layer->map);
-    layer->unmap.notify = handle_layer_unmap;
-    wl_signal_add(&layer_surface->events.unmap, &layer->unmap);
-    layer->destroy.notify = handle_layer_destroy;
-    wl_signal_add(&layer_surface->events.destroy, &layer->destroy);
-    layer->surface_commit.notify = handle_layer_surface_commit;
-    wl_signal_add(&layer_surface->surface->events.commit, &layer->surface_commit);
 
     // Temporarily set the layer's current state to pending
     // So that we can easily arrange it
